third_party/blink/web_tests/virtual/cross-partition-blob-url/external/wpt/storage-access-api/storage-access-beyond-cookies.BlobURLDedicatedWorker.sub.https.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify that if the third-party context creates a blob URL using the StorageAccessHandle and sends it to the dedicated worker, the dedicated worker fetch succeeds.
  assert_equals: expected "Blob URL DedicatedWorker tests completed successfully." but got "Dedicated worker unexpectedly fetched first-party blob URL from a third-party context without granting storage access."
Harness: the test ran to completion.
